requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg[binary]==3.2.3
passlib==1.7.4
bcrypt==4.1.3
pydantic[email]==2.7.1
python-multipart==0.0.9
boto3==1.34.147
python-magic==0.4.27
